Fixed a warning

+++ finding.cpp
@@ -5,7 +5,7 @@ using namespace std;
 /***********************FUNCTION DEFINITIONS***********************/
 
 int findCriminal(string array[], int SIZE){
-    int min = array[0].length();
+    long min = array[0].length();
     int returnValue = 0;
     for (int i=1; i<SIZE; i++){
         if(array[i].length() > min){
